pixie 0.7.15 (new formula)
This adds a homebrew based installation for the Pixie CLI.

Pixie is an open-source observability tool for Kubernetes applications,
and the CLI is used to deploy Pixie, manage it, and interact with the
platform (for e.g. run queries).

Pixie already provides release binaries and install scripts but some folks
prefer to manage all their tools using homebrew so this gives folks that option.

Fixes https://github.com/pixie-io/pixie/issues/401

Closes #104328.

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>
Signed-off-by: Rui Chen <rui@chenrui.dev>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>