crystal: add interpreter
Crystal includes an interpreter. By default it doesn't come with it,
you have to explicitly pass `interpreter=true` when compiling Crystal.

The reason it doesn't come by default with it is that it depends
on libffi, and it also won't work with statically distributed versions
of Crystal.

However, on Mac including the interpreter, especially via Homebrew,
should be very easy to do.

I think it makes sense to include the interpreter by default,
specially becuase in the upcoming 1.5.0 release it will be much
more stable.

Also, while we're here, let's do a little clean up of the formula and
set `CRYSTAL_CONFIG_LIBRARY_PATH` so that we end up with a relocatable
binary.

Co-authored-by: Carlo Cabrera <30379873+carlocab@users.noreply.github.com>