teleport: enable libfido2 features
Teleport v10 added support for passwordless, which requires the `tsh`
binary to link against libfido2. This change adds the libfido2
dependency to the formula and the necessary make switch to enable
dynamic linking against it.

The make switch is a noop for versions <= v10.0.2, but ready to land on
v10.0.3 or v10.1.0 (whatever releases first). Because of that, it's not
necessary to bump the revision or rebuild the bottles.

Closes #106359.

Signed-off-by: Sean Molenaar <1484494+SMillerDev@users.noreply.github.com>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>