luajit: remove `inreplace` for `LUA_ROOT`
We can achieve the same effect of the `inreplace` without actually using
an `inreplace` by building luajit with `PREFIX=$HOMEBREW_PREFIX`. We
make sure it's installed to the right place by passing the real prefix
when we do `make install`.

This allows us to stop relying on a complicated and fragile `inreplace`.

Additionally, install the docs provided with the source tarball to avoid
any mismatches between the online documentation and the version we
actually install.

Closes #107290.

Signed-off-by: Sean Molenaar <1484494+SMillerDev@users.noreply.github.com>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>