libvatek 3.04 (new formula)
The Vatek C/C++ library is used by TV applications to control modulator
devices based on Vatek chips. Because these chips are supported on macOS,
this makes them the first (and currently only) TV modulators available
on macOS. Other existing modulators are available on Linux and Windows
only.

This is a new library which is provided as open source by Vatek, the
company which produces these chips.

This library will be used by the next version of TSDuck. This library
will be consequently required as a dependency for the tsduck formula
in its next version.

As a committer, I am not the owner of that library. However, the
development team of libvatek gave me the permission to initiate the
first commit into Homebrew with that name and version. They are not
yet familiar with Homebrew.

Please note that brew audit --new-formula libvatek reports this error:
 * GitHub repository not notable enough (<30 forks, <30 watchers and <75 stars)

Please ignore this error and allow this formula into Homebrew. New libraries
must start somewhere. They become notable when they are used. And users
need packages to use a library. Additionally, not allowing this library
into Homebrew would block the next version of the tsduck formula.

Thanks.

Closes #107888.

Signed-off-by: Carlo Cabrera <30379873+carlocab@users.noreply.github.com>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>