tbb: apply patch to fix thread creation under heavy load
TBB has a bug in thread creation which is exposed with usage in mold.
See rui314/mold#410, rui314/mold#600. We apply a patch that has been
upstreamed to TBB at oneapi-src/oneTBB#824 to fix this.

The patch implements a mechanism similar to one used by Go [1], and has
been adopted in the Arch and OpenSUSE TBB packages.

While we're here, let's align the `cmake` invocation with other
formulae, fix references to `python3.10` (#108008), and change the
`-rpath` flag so that it does not require relocation when pouring a
bottle (which should speed bottle pour times up slightly).

[1] https://go-review.googlesource.com/c/go/+/33894/

Closes #108431.

Signed-off-by: Sean Molenaar <1484494+SMillerDev@users.noreply.github.com>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>