libfaketime: fix tests for ARM macOS (#109894)
* libfaketime: Fix tests for macOS Monterey

The workaround previously made of copying `/bin/date` to a different location
doesn't work anymore since macOS Monterey (12).

This fixes that by instead compiling a small C program that emits the current
time as a Unix timestamp, making it easy to compare against the expected
output.

* Minor formatting fixes, suggested by @carlocab

* Add the TZ environment variable back

The macOS arm64 instances are located in UTC+8 which throws off the tests. This
might correct that.