mold: build a relocatable bottle.
The build defines a `LIBDIR` macro which embeds the `lib` directory in
the `mold` binary.

If we undefine this macro, `mold` will fall back to looking for the
`lib` directory relative to itself, so it should continue to function
correctly even without this. This has the benefit of making the bottle
relocatable.

Our test should exercise this code path, so any breakage will be
detected in CI.

Closes #111967.

Signed-off-by: Sean Molenaar <1484494+SMillerDev@users.noreply.github.com>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>