re2: build and install static library
e07cb240d6232e42223ab29b7433e23de4ee4801 switched to using cmake, but
cmake only builds the shared (.so) or static library (.a) depending on
whether `BUILD_SHARED_LIBS` is enabled.

To fix this, we run `cmake` twice: once to build the static library,
and once to build the shared library.

Closes #119928.

Signed-off-by: Rui Chen <rui@chenrui.dev>
Signed-off-by: BrewTestBot <1589480+BrewTestBot@users.noreply.github.com>