vim: Add libsodium support and fix `brew test vim`
Vim added libsodium a while ago for handling reading/writing encrypted
files, and we should support it in the Homebrew distribution. Also added
checks to make sure `+sodium` is in the version output.

Also, fix `brew test vim` not working on a local machine, since it
cannot find gettext properly. To properly instruct configure to look
under the prefix (e.g. /opt/homebrew), we actually need to pass
`--with-local-dir` to the configure script, as otherwise it's going to
use /usr/local. I think this may be why it's working in build systems
but not my local machine. Note that macvim.rb is already setting this
flag in the configure script and therefore works fine.