macvim: Add libsodium support and use build-time-only deps
Vim added libsodium a while ago for handling reading/writing encrypted
files, and we should support it in the Homebrew distribution for MacVim.
Also added checks to make sure `+sodium` is in the version output.

Also, MacVim has been building with statically linked libraries (gettext
and sodium) in its build system, unlike Vim which uses dynamic linking
in its configure script. Because of that, we don't actually have a
runtime dependency on those libraries. Move them to build-time
dependnecy instead.