tests: re-check labels for dependent testing
The change from #128809 doesn't work as expected, because `github.event`
doesn't seem to pick up changes in labels at some point after the PR is
opened. See #128865 for an example.

In order to be able to change the labels and have them apply only to
dependent tests, let's check the labels again in a separate job for
dependent testing. In order to minimise code repitition, we extract the
bulk of the `check-labels` job into a separate script.