remove-long-timeout-labels: fix label removal when CI is restarted
Restarting CI (e.g. from a force push) causes this workflow to remove
the long-timeout label unexpectedly.

Let's fix that by waiting briefly and then checking the PR's CI status
to make sure that we're not removing the long timeout label after CI has
been restarted.

To avoid having to run this workflow for another minute or so only to
end up not doing anything (because the PR did not have the long timeout
label), let's check that the PR does in fact have the relevant label
before proceeding.