workflows: add `create-replacement-pr.yml`
This adds a workflow that can be dispatched to replace an existing pull
request.

I intend for this workflow to be dispatched automatically when we have
no push access to a PR, or when a PR needs to be autosquashed. Using
this to autosquash PRs will help us avoid errors due to insufficient
scopes in BrewTestBot's PAT.