remove-long-timeout-labels: relax requirements
We currently require CI to complete before a long timeout label can be
removed from a PR.

Let's relax that a little by allowing removal of the label if at most
one macOS dependent job is still running. In particular, this allows
removal of the label if the Linux runner is the only remaining job still
running.