triage-ci: show new contributor message only once
Our new contributor message contains a workflow run ID, which generally
changes across CI runs. This means that our check for having shown this
message will always fail.

Let's fix that by showing a message that always stays the same so it's
easier to check the previous comments for whether we've posted it
previously.