sox: build with alsa-lib on linux
sox with alsa on linux (update formula)

Up to now, on linux, sox has been built with support for oss only,
which however is no longer the standard for modern linux
systems. Therefore invoking play (e.g.) fails with an error. This
patch adds build option '--with-alsa' to build sox with support for
alsa in addition to oss. As a result play actually produces sound.

Sox: Add License information GPL-2.0

Source are the files COPYING and LICENSE.GPL in the top level of sox
main directory; note that there is no mention of "later version"
there.

sox: changes License to GPL-2.0-only

As the files COPYING and LICENSE.GPL only speak of GPL 2 and
do not mention later versions, this needs to be tranlated in
GPL-2.0-only rather than GPL-2.0.

sox: add web-source for given license

As sourceforge is the code-repository, the License stated there
should be used.

sox: add LGPL to license clause

Made the license-clause a compound to include LGPL too; more precise
instructions to find the license on the project homepage.

sox: removed comments about license

The discussion happened in https://github.com/Homebrew/homebrew-core/pull/136129
and cn be viewed there, if need be. No need to clutter the formula.

sox: update build

Co-authored-by: Sean Molenaar <SMillerDev@users.noreply.github.com>